python3: fix incorrect use of merging two dictionaries
In python2, this was inefficient, because it allocated *three* lists of
tuples, before finally generating a dict based on them. In python3, it
fails because you cannot combine the dict_items() type.

Moreover, retval was always a function-local dictionary used purely for
returning the value, so dict1.update(dict2) will always yield the
correct result, and we don't even need to create a copy to avoid
mutating the original dictionary.